Refactor report router into fetchData function
